refactor: make document list view use containers
