+++ frontend/components/DocumentListToolbar/DocumentListToolbar.css
@@ -1,0 +1,3 @@
+.count-label {
+  margin-left: 15px;
+}
