+++ frontend/containers/App/App.jsx
@@ -13,7 +13,7 @@ import LoadingBar from 'components/LoadingBar/LoadingBar'
 import Main from 'components/Main/Main'
 
 import DocumentEdit from 'views/DocumentEdit/DocumentEdit'
-import DocumentList from 'views/DocumentList/DocumentList'
+import DocumentListView from 'views/DocumentListView/DocumentListView'
 import Error from 'views/Error/Error'
 import Home from 'views/Home/Home'
 import MediaLibrary from 'views/MediaLibrary/MediaLibrary'
@@ -107,8 +107,8 @@ class App extends Component {
             <PasswordReset path="/reset" authenticate/>
             {hasRoutes && ( <DocumentEdit path="/:group/:collection/document/:method/:documentId?/:section?" authenticate /> )}
             <DocumentEdit path="/:collection/document/:method/:documentId?/:section?" authenticate />
-            <DocumentList path="/:collection/documents/:page?" authenticate />
-            {hasRoutes && ( <DocumentList path="/:group/:collection/documents/:page?" authenticate /> )}
+            <DocumentListView path="/:collection/documents/:page?" authenticate />
+            {hasRoutes && ( <DocumentListView path="/:group/:collection/documents/:page?" authenticate /> )}
             <MediaLibrary path="/:collection/media/:document?" authenticate/>
             {hasRoutes && ( <MediaLibrary path="/:group/:collection/media/:document?" authenticate/> )}
             <UserProfile path="/profile" authenticate />
